voucher is move to type guard
