+++ src/interfaces/voucher.ts
@@ -1,9 +1,6 @@
 import { VOUCHER_TYPE } from '@/constants/account';
-import { isVoucherType } from '@/libs/utils/type_guard/account';
 import { ILineItem } from '@/interfaces/line_item';
 import { IPayment } from '@/interfaces/payment';
-import { isIPayment } from '@/libs/utils/type_guard/payment';
-import { isILineItem } from '@/libs/utils/type_guard/line_item';
 
 export interface IVoucherMetaData {
   date: number;
@@ -11,7 +8,7 @@ export interface IVoucherMetaData {
   companyId: string;
   companyName: string;
   description: string;
-  reason: string; // 從paymentReason改這個
+  reason: string;
   projectId: string;
   project: string;
   contractId: string;
@@ -24,38 +21,4 @@ export interface IVoucher {
   invoiceIndex: string;
   metadatas: IVoucherMetaData[];
   lineItems: ILineItem[];
-}
-
-function isIVoucherMetaData(arg: IVoucherMetaData): arg is IVoucherMetaData {
-  if (
-    typeof arg.date !== 'number' ||
-    isVoucherType(arg.VOUCHER_TYPE) ||
-    typeof arg.companyId !== 'string' ||
-    typeof arg.companyName !== 'string' ||
-    typeof arg.description !== 'string' ||
-    typeof arg.reason !== 'string' ||
-    typeof arg.projectId !== 'string' ||
-    typeof arg.project !== 'string' ||
-    typeof arg.contractId !== 'string' ||
-    typeof arg.contract !== 'string' ||
-    isIPayment(arg.payment)
-  ) {
-    return false;
-  }
-  return true;
-}
-
-export function isIVoucher(arg: IVoucher): arg is IVoucher {
-  if (arg.voucherIndex === undefined || arg.invoiceIndex === undefined) {
-    return false;
-  }
-  const isIVoucherMetaDataReturn = arg.metadatas.every(isIVoucherMetaData);
-  if (isIVoucherMetaDataReturn) {
-    return false;
-  }
-  const isILineItemReturn = arg.lineItems.every(isILineItem);
-  if (!isILineItemReturn) {
-    return false;
-  }
-  return true;
 }
